backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
python-dotenv==1.0.1
python-multipart==0.0.20
anthropic==0.42.0
python-docx==1.1.2
reportlab==4.2.5
PyMuPDF==1.25.3
bcrypt==4.2.1
PyJWT==2.10.1
